Move entries mock data to a json file and improve structure
