+++ app/models/__init__.py
@@ -1,2 +1,1 @@
-from mock import Mock
 from entry import ModelNotFoundException, Entry
