+++ utils.py
@@ -1,0 +1,7 @@
+import os
+
+ROOT_DIRECTORY = os.path.dirname(__file__)
+
+
+def full_path(path):
+    return ROOT_DIRECTORY + '/' + path
